pyproject.toml
[project]
name = "klausulaai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "python-dotenv>=1.2.2",
    "supabase>=2.29.0",
    "uvicorn>=0.46.0",
]
